test del roomQuestion completados
